Fixed MSVC compiler error C3872 by replacing the Greek alphabet symbols with their corresponding English names.
Compiling the current code as it is using MSVC 2022 throws a lot of C3872 compiler errors such as "error C3872: '0x2030': this character is not allowed in an identifier".
This is due to the fact that the MSVC compiler 'follows the C++11 standard on characters allowed in an identifier'. See https://learn.microsoft.com/en-us/cpp/error-messages/compiler-errors-2/compiler-error-c3872?view=msvc-170.
The issue is resolved by replacing the Greek alphabet symbols with their corresponding English names.